requirements.txt
Django==1.7.4
django-admin-honeypot==0.4.0
django-appconf==0.6
django-compressor==1.4
django-debug-toolbar==1.2.2
django-devserver==0.8.0
django-extensions==1.4.9
django-grappelli==2.6.3
django-imagekit==3.2.5
django-model-utils==2.2
django-querycount==0.1.1
django-secure==1.0.1
djangorestframework==3.0.5
guppy==0.1.10
ipdb==0.8
ipython==2.3.1
line-profiler==1.0
pilkit==1.1.12
Pillow==2.7.0
psycopg2==2.5.4
six==1.9.0
sqlparse==0.1.14
Werkzeug==0.9.6
